Advent of Code 2021, Day 6
Day 6 of the Advent of Code 2021 set of puzzles. This one was
a little challenging as the first half puzzle wasn't obvious in
what the second half would be. Accounting for that, the solution
used for the first half of the puzzle wasn't expandable to the
second.

So a bit of a rewrite was in order for part two of the puzzle
and that's reflected in the bottom of the Day06.clj file.

https://adventofcode.com/2021/day/6